[nmpc][py]: move unused file to the archive folder
