+++ aerial_robot_control/scripts/nmpc/tilt_qd/archive/tilt_qd_servo_thrust_drag.py
