+++ aerial_robot_control/scripts/nmpc/tilt_qd/archive/tilt_qd_servo_w_cog_end_dist.py
